score needs to be double
see 2e9426c

+++ lima_linguisticprocessing/src/linguisticProcessing/core/PosTagger/DynamicSvmToolPosTagger.h
@@ -96,7 +96,7 @@ private:
    */
   struct PathInfo {
     LinguisticGraphVertex prev;
-    uint64_t score;
+    double score;
     std::string pos;
     int pathLength;
   };
